# backend/openapi.yaml
openapi: 3.1.0
info:
  title: "KAVACH \u2014 AI-Powered DevSecOps Platform"
  description: "KAVACH scans banking application repositories for vulnerabilities,\
    \ calculates a Banking Risk Score (BRS), maps findings to regulatory controls\
    \ (RBI / PCI-DSS / SWIFT), and generates AI-powered audit reports.\n\n## Pipeline\n\
    `GitHub webhook` \u2192 `Create Scan Job` \u2192 `Queue` \u2192 `Distributed workers`\
    \ \u2192 `Aggregation` \u2192 `Risk Engine` \u2192 `Compliance Engine` \u2192\
    \ `AI Explanation` \u2192 `Reports` \u2192 `Dashboard` \u2192 `Notifications`\
    \ \u2192 `Archive`\n\nA scan enters this pipeline either via `POST /scan/repository`\
    \ / `POST /scan` (an authenticated user submitting a URL or archive) or `POST\
    \ /webhooks/github` (an unauthenticated, HMAC-verified push event) \u2014 both\
    \ converge on the same queue \u2192 distributed-worker \u2192 aggregation path,\
    \ so everything downstream behaves identically regardless of how the scan was\
    \ triggered."
  contact:
    name: KAVACH Platform Team
    url: https://github.com/kavach-platform/kavach
  license:
    name: "Proprietary \u2014 All Rights Reserved"
  version: 1.0.0
paths:
...
  /api/v1/auth/register:
    post:
      tags:
      - Auth
      summary: Register
      description: "Create a new account (always the least-privileged role \u2014\
        \ see UserRegisterRequest). Raises 409 if the email is already registered."
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - Auth
      summary: Login
      description: "OAuth2 password flow \u2014 `username` field carries the email."
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_login_api_v1_auth_login_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/refresh:
    post:
      tags:
      - Auth
      summary: Refresh Token
      operationId: refresh_token_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - Auth
      summary: Get Me
      operationId: get_me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/auth/sso/oauth2/callback:
    get:
      tags:
      - "Auth \u2014 SSO"
      summary: Oauth2 Callback
      operationId: oauth2_callback_api_v1_auth_sso_oauth2_callback_get
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
          title: Code
      - name: state
        in: query
        required: true
        schema:
          type: string
          title: State
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/auth/ldap/login:
    post:
      tags:
      - "Auth \u2014 SSO"
      summary: Ldap Login
      operationId: ldap_login_api_v1_auth_ldap_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LDAPLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/admin/users:
    get:
      tags:
      - "Auth \u2014 Admin"
      summary: List Users
      operationId: list_users_api_v1_auth_admin_users_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserRead'
                title: Response List Users Api V1 Auth Admin Users Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/admin/users/{user_id}/role:
    patch:
      tags:
      - "Auth \u2014 Admin"
      summary: Update User Role
      operationId: update_user_role_api_v1_auth_admin_users__user_id__role_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/admin/users/{user_id}/active:
    patch:
      tags:
      - "Auth \u2014 Admin"
      summary: Update User Active Status
      operationId: update_user_active_status_api_v1_auth_admin_users__user_id__active_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActiveStatusUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/audit-log:
    get:
      tags:
      - "Auth \u2014 Admin"
      summary: Query Audit Log
      description: 'Defaults to the last `audit_log_default_lookback_days` days (90
        by

        default) when `since` isn''t given, so a careless unbounded query can''t

        accidentally full-scan the entire table.'
      operationId: query_audit_log_api_v1_auth_audit_log_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: User Id
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Since
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ActiveStatusUpdateRequest:
      properties:
        is_active:
          type: boolean
          title: Is Active
      type: object
      required:
      - is_active
      title: ActiveStatusUpdateRequest
    AuditLogEntryResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: User Id
        user_email:
          anyOf:
          - type: string
          - type: 'null'
          title: User Email
        action:
          type: string
          title: Action
        resource_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Type
        resource_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Id
        status:
          type: string
          title: Status
        ip_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Ip Address
        user_agent:
          anyOf:
          - type: string
          - type: 'null'
          title: User Agent
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Details
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - action
      - status
      - created_at
      title: AuditLogEntryResponse
    AuditLogListResponse:
      properties:
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
        entries:
          items:
            $ref: '#/components/schemas/AuditLogEntryResponse'
          type: array
          title: Entries
      type: object
      required:
      - total
      - limit
      - offset
      - entries
      title: AuditLogListResponse
    AuthProvider:
      type: string
      enum:
      - local
      - oauth2
      - saml
      - ldap
      title: AuthProvider
      description: "How a user's identity was established \u2014 see app/auth/sso/."
...
    Body_login_api_v1_auth_login_post:
      properties:
        grant_type:
          anyOf:
          - type: string
            pattern: ^password$
          - type: 'null'
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          format: password
          title: Password
        scope:
          type: string
          title: Scope
          default: ''
        client_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
          - type: string
          - type: 'null'
          format: password
          title: Client Secret
      type: object
      required:
      - username
      - password
      title: Body_login_api_v1_auth_login_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LDAPLoginRequest:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
      type: object
      required:
      - username
      - password
      title: LDAPLoginRequest
...
    RefreshTokenRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshTokenRequest
...
    RoleUpdateRequest:
      properties:
        role:
          $ref: '#/components/schemas/UserRole'
      type: object
      required:
      - role
      title: RoleUpdateRequest
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
      type: object
      required:
      - access_token
      - refresh_token
      title: TokenResponse
    UserRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          format: email
          title: Email
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          title: Is Active
        auth_provider:
          $ref: '#/components/schemas/AuthProvider'
        role_display_name:
          type: string
          title: Role Display Name
          readOnly: true
        permissions:
          items:
            type: string
          type: array
          title: Permissions
          readOnly: true
      type: object
      required:
      - id
      - email
      - role
      - is_active
      - auth_provider
      - role_display_name
      - permissions
      title: UserRead
    UserRegisterRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 128
          minLength: 8
          title: Password
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
      type: object
      required:
      - email
      - password
      title: UserRegisterRequest
    UserRole:
      type: string
      enum:
      - admin
      - auditor
      - developer
      - security_engineer
      - read_only
      title: UserRole
      description: "The fixed RBAC role set \u2014 see app/auth/permissions.py for\
        \ what each\nrole can actually do (ROLE_PERMISSIONS). Not user-extensible:\
        \ adding a\nrole means adding an enum member + a migration, the same as any\
        \ other\nenum in this file, rather than a dynamic Role table \u2014 there\
        \ are 5 of\nthese by design, not 5000."
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
...